Update to latest version of tinygo and KW's Go SDK
Latest version of TinyGo has all the fixes we need. We can now move away
from our "hack".

As part of the transition, we also have to update to latest stable
release of KW's Go SDK. This version provides a different way to mock
the waPC client, hence some changes had to be done to our codebase.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>